timestamp -> Double Timestamp to ChallengeCardView and GeneralView added

+++ talento/Data/ChallengeData.swift
@@ -47,7 +47,7 @@ class ChallengeData: ObservableObject{
                             image: document.get("image") as! String,
                             participants: document.get("participants") as! [User.ID],
                             title: document.get("title") as! String,
-                            timestamp: document.get("timestamp") as! Int,
+                            timestamp: document.get("timestamp") as! Double,
                             duration: document.get("duration") as? Double ?? 1,
                             instructions: document.get("instructions") as! String,
                             isPublic: document.get("isPublic") as? Bool,
